requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
django-extensions==3.0.8
django-redis-cache==3.0.0
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
PyJWT==1.7.1
coverage==5.3
freezegun==1.0.0
whitenoise==5.2.0
beautifulsoup4==4.9.3
